experiments/mcp-prompt-injection-corpus/requirements.txt
pyyaml==6.0.1
pytest==9.0.3
